@19622 	Modified the DDL writer to check if a create script already 	exists for a given table and not overwrite it if it does. This 	will allow us to incrementally replace manually specified 	create tables with auto generated ones.

+++ archive/core-platform/src/com/arsdigita/persistence/metadata/DDLWriter.java
@@ -9,21 +9,31 @@ import java.util.*;
  * DDLWriter
  *
  * @author <a href="mailto:rhs@mit.edu">rhs@mit.edu</a>
- * @version $Revision: #1 $ $Date: 2002/08/06 $
+ * @version $Revision: #2 $ $Date: 2002/08/06 $
  **/
 
 public class DDLWriter {
 
-    public final static String versionId = "$Id: //core-platform/dev/src/com/arsdigita/persistence/metadata/DDLWriter.java#1 $ by $Author: rhs $, $DateTime: 2002/08/06 16:54:58 $";
+    public final static String versionId = "$Id: //core-platform/dev/src/com/arsdigita/persistence/metadata/DDLWriter.java#2 $ by $Author: rhs $, $DateTime: 2002/08/06 18:26:49 $";
 
     private File m_base;
+    private boolean m_overwrite;
 
     public DDLWriter(String base) {
-        this(new File(base));
+        this(new File(base), false);
+    }
+
+    public DDLWriter(String base, boolean overwrite) {
+        this(new File(base), overwrite);
     }
 
     public DDLWriter(File base) {
+        this(base, false);
+    }
+
+    public DDLWriter(File base, boolean overwrite) {
         m_base = base;
+        m_overwrite = overwrite;
         if (!m_base.isDirectory()) {
             throw new IllegalArgumentException("expecting directory");
         }
@@ -35,10 +45,20 @@ public class DDLWriter {
 
     public void write(Collection tables)
         throws IOException {
+        Set skipped = new HashSet();
+
         for (Iterator it = tables.iterator(); it.hasNext(); ) {
             Table table = (Table) it.next();
 
-            File file = new File(m_base, table.getName() + "-create.sql");
+            File file = new File(m_base, "table-" + table.getName() +
+                                 ".sql");
+            if (!m_overwrite && file.exists()) {
+                skipped.add(table);
+                continue;
+            }
+
+            file = new File(m_base, "table-" + table.getName() + "-auto.sql");
+
             FileWriter writer = new FileWriter(file);
             writer.write(table.getSQL());
             writer.write("\n");
@@ -82,6 +102,9 @@ public class DDLWriter {
                                                         "deffered.sql"));
             for (Iterator it = deffered.iterator(); it.hasNext(); ) {
                 Table table = (Table) it.next();
+                if (skipped.contains(table)) {
+                    continue;
+                }
                 for (Iterator iter = table.getConstraints().iterator();
                      iter.hasNext(); ) {
                     Constraint con = (Constraint) iter.next();
@@ -99,7 +122,11 @@ public class DDLWriter {
         FileWriter writer = new FileWriter(new File(m_base, "create.sql"));
         for (Iterator it = createOrder.iterator(); it.hasNext(); ) {
             Table table = (Table) it.next();
-            writer.write("@@" + table.getName() + "-create.sql\n");
+            if (skipped.contains(table)) {
+                writer.write("@@table-" + table.getName() + ".sql\n");
+            } else {
+                writer.write("@@table-" + table.getName() + "-auto.sql\n");
+            }
         }
         if (deffered.size() > 0) {
             writer.write("@@deffered.sql\n");
